chore: update dependencies
Update dependencies to the latest version. This includes an update of
`eslint-config-typescript-shareable` that makes a few import/* rules
function as intended again.

Most of the changes in this commit are related to that update. Most are
simple import changes from `.` to `./Component`. On a few occasions,
some parts were refactored out.