feat(button): danger and success variant
Adds a "danger" variant which signifies a dangerous
action, such as deleting a user.

Adds a "success" variant which signifies a positive
action.

The change affects

- Button
- IconButton
- IconTextButton

This change also affects a number of snapshots. This because
a primary button uses `textPrimary` and `elementPrimary` to
change the color between for instance the normal and hover state.
`textError` and `elementError` however had the exact same color.

Had to add `textSuccess` since this color didn't exist at all,
which seems like an error due to the fact that all other `element...`
has an equivalent `text...`.