feat(tabs)!: add tabs component
Add a new Tabs and VerticalTabs (plural)
component and removes the Tab and VerticalTab
components.
I does not make sense to give the user the ability
to use Tabs individually, so instead we created
a wrapper component which takes the tabs as options
instead.

This will also make our PRC => Fluent conversion
easier because Fluent has a wrapper component
that handles all the clicking and switching logic.

BREAKING CHANGE: replaces Tab and VerticalTab components
with Tabs and VerticalTabs components.